+++ src/components/card/card.stories.ts
@@ -69,20 +69,18 @@ export const FooterLinks = (): string => html`
 
 export const FooterTextButtonsTooltips = (): string => html`
   <div style="width:260px">
-    <calcite-tooltip-manager>
-      <calcite-card ${boolean("loading", false)} ${boolean("selectable", false)}>
-        ${footerThumbnail}
-        <h3 slot="title">My great project that might wrap two lines</h3>
-        <span slot="subtitle">Johnathan Smith</span>
-        <span slot="footer-leading">Nov 25, 2018</span>
-        <div slot="footer-trailing">
-          <calcite-button id="card-icon-test-6" scale="s" appearance="transparent" color="neutral" icon-start="circle">
-          </calcite-button>
-          <calcite-button id="card-icon-test-7" scale="s" appearance="transparent" color="neutral" icon-start="circle">
-          </calcite-button>
-        </div>
-      </calcite-card>
-    </calcite-tooltip-manager>
+    <calcite-card ${boolean("loading", false)} ${boolean("selectable", false)}>
+      ${footerThumbnail}
+      <h3 slot="title">My great project that might wrap two lines</h3>
+      <span slot="subtitle">Johnathan Smith</span>
+      <span slot="footer-leading">Nov 25, 2018</span>
+      <div slot="footer-trailing">
+        <calcite-button id="card-icon-test-6" scale="s" appearance="transparent" color="neutral" icon-start="circle">
+        </calcite-button>
+        <calcite-button id="card-icon-test-7" scale="s" appearance="transparent" color="neutral" icon-start="circle">
+        </calcite-button>
+      </div>
+    </calcite-card>
   </div>
   <calcite-tooltip placement="top-start" reference-element="card-icon-test-6">Configure </calcite-tooltip>
   <calcite-tooltip placement="bottom-start" reference-element="card-icon-test-7">Delete </calcite-tooltip>
@@ -92,7 +90,6 @@ FooterTextButtonsTooltips.storyName = "Footer Text, Buttons, Tooltips";
 
 export const FooterButtonsTooltipsDropdown = (): string => html`
 <div style="width:260px">
-  <calcite-tooltip-manager>
     <calcite-card
     ${boolean("loading", false)}
     ${boolean("selectable", false)}
@@ -126,7 +123,6 @@ export const FooterButtonsTooltipsDropdown = (): string => html`
         </calcite-dropdown>
       </div>
     </calcite-card>
-  </calcite-tooltip-manager>
 </div>
 <calcite-tooltip placement="bottom-start" reference-element="card-icon-test-1">My great tooltip example
 </calcite-tooltip>
@@ -203,20 +199,18 @@ DarkThemeFooterLinks.parameters = { themes: themesDarkDefault };
 
 export const DarkThemeFooterTextButtonsTooltips = (): string => html`
   <div style="width:260px">
-    <calcite-tooltip-manager>
-      <calcite-card class="calcite-theme-dark" ${boolean("loading", false)} ${boolean("selectable", false)}>
-        ${footerThumbnail}
-        <h3 slot="title">My great project that might wrap two lines</h3>
-        <span slot="subtitle">Johnathan Smith</span>
-        <span slot="footer-leading">Nov 25, 2018</span>
-        <div slot="footer-trailing">
-          <calcite-button id="card-icon-test-6" scale="s" appearance="transparent" color="neutral" icon-start="circle">
-          </calcite-button>
-          <calcite-button id="card-icon-test-7" scale="s" appearance="transparent" color="neutral" icon-start="circle">
-          </calcite-button>
-        </div>
-      </calcite-card>
-    </calcite-tooltip-manager>
+    <calcite-card class="calcite-theme-dark" ${boolean("loading", false)} ${boolean("selectable", false)}>
+      ${footerThumbnail}
+      <h3 slot="title">My great project that might wrap two lines</h3>
+      <span slot="subtitle">Johnathan Smith</span>
+      <span slot="footer-leading">Nov 25, 2018</span>
+      <div slot="footer-trailing">
+        <calcite-button id="card-icon-test-6" scale="s" appearance="transparent" color="neutral" icon-start="circle">
+        </calcite-button>
+        <calcite-button id="card-icon-test-7" scale="s" appearance="transparent" color="neutral" icon-start="circle">
+        </calcite-button>
+      </div>
+    </calcite-card>
   </div>
   <calcite-tooltip placement="bottom" class="calcite-theme-dark" reference-element="card-icon-test-6"
     >Configure
@@ -231,66 +225,64 @@ DarkThemeFooterTextButtonsTooltips.parameters = { themes: themesDarkDefault };
 
 export const DarkThemeFooterButtonsTooltipsDropdown = (): string => html`
   <div style="width:260px">
-    <calcite-tooltip-manager>
-      <calcite-card class="calcite-theme-dark" ${boolean("loading", false)} ${boolean("selectable", false)}>
-        <img
-          alt=""
-          slot="thumbnail"
-          src="${placeholderImage({ width: 260, height: 160 })}"
-          style="width: 260px; height: 160px;"
-        />
-        <h3 slot="title">Portland Businesses</h3>
-        <span slot="subtitle"
-          >by
-          <calcite-link href="">example_user</calcite-link>
-        </span>
-        <div>
-          Created: Apr 22, 2019
-          <br />
-          Updated: Dec 9, 2019
-          <br />
-          View Count: 0
-        </div>
+    <calcite-card class="calcite-theme-dark" ${boolean("loading", false)} ${boolean("selectable", false)}>
+      <img
+        alt=""
+        slot="thumbnail"
+        src="${placeholderImage({ width: 260, height: 160 })}"
+        style="width: 260px; height: 160px;"
+      />
+      <h3 slot="title">Portland Businesses</h3>
+      <span slot="subtitle"
+        >by
+        <calcite-link href="">example_user</calcite-link>
+      </span>
+      <div>
+        Created: Apr 22, 2019
+        <br />
+        Updated: Dec 9, 2019
+        <br />
+        View Count: 0
+      </div>
+      <calcite-button
+        slot="footer-leading"
+        color="neutral"
+        scale="s"
+        id="card-icon-test-1"
+        icon-start="circle"
+      ></calcite-button>
+      <div slot="footer-trailing">
         <calcite-button
-          slot="footer-leading"
+          class="calcite-theme-dark"
           color="neutral"
           scale="s"
-          id="card-icon-test-1"
+          id="card-icon-test-2"
           icon-start="circle"
         ></calcite-button>
-        <div slot="footer-trailing">
+        <calcite-button
+          class="calcite-theme-dark"
+          color="neutral"
+          scale="s"
+          id="card-icon-test-3"
+          icon-start="circle"
+        ></calcite-button>
+        <calcite-dropdown type="hover">
           <calcite-button
             class="calcite-theme-dark"
             color="neutral"
+            id="card-icon-test-5"
+            slot="dropdown-trigger"
             scale="s"
-            id="card-icon-test-2"
             icon-start="circle"
           ></calcite-button>
-          <calcite-button
-            class="calcite-theme-dark"
-            color="neutral"
-            scale="s"
-            id="card-icon-test-3"
-            icon-start="circle"
-          ></calcite-button>
-          <calcite-dropdown type="hover">
-            <calcite-button
-              class="calcite-theme-dark"
-              color="neutral"
-              id="card-icon-test-5"
-              slot="dropdown-trigger"
-              scale="s"
-              icon-start="circle"
-            ></calcite-button>
-            <calcite-dropdown-group selection-mode="none">
-              <calcite-dropdown-item>View details</calcite-dropdown-item>
-              <calcite-dropdown-item>Duplicate</calcite-dropdown-item>
-              <calcite-dropdown-item>Delete</calcite-dropdown-item>
-            </calcite-dropdown-group>
-          </calcite-dropdown>
-        </div>
-      </calcite-card>
-    </calcite-tooltip-manager>
+          <calcite-dropdown-group selection-mode="none">
+            <calcite-dropdown-item>View details</calcite-dropdown-item>
+            <calcite-dropdown-item>Duplicate</calcite-dropdown-item>
+            <calcite-dropdown-item>Delete</calcite-dropdown-item>
+          </calcite-dropdown-group>
+        </calcite-dropdown>
+      </div>
+    </calcite-card>
   </div>
   <calcite-tooltip
     class="calcite-theme-dark"
