+++ src/components/tooltip/usage/Basic.md
@@ -2,10 +2,8 @@
 <calcite-tooltip placement="auto" reference-element="tooltip-button"
   >This is the message of the tooltip</calcite-tooltip
 >
-<calcite-tooltip-manager>
-  <p>
-    Lorem <a id="tooltip-button" href="#">ipsum</a> dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-    incididunt ut labore et dolore magna aliqua.
-  </p>
-</calcite-tooltip-manager>
+<p>
+  Lorem <a id="tooltip-button" href="#">ipsum</a> dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
+  incididunt ut labore et dolore magna aliqua.
+</p>
 ```
