+++ src/demos/card.html
@@ -76,69 +76,58 @@
       <!-- Normal card -->
       <div class="child">
         <p>Card with several buttons</p>
-        <calcite-tooltip-manager>
-          <calcite-card>
-            <img slot="thumbnail" alt="Sample image alt" src="https://placem.at/places?w=260&h=160&txt=0" />
-            <h3 slot="title">Portland Businesses</h3>
-            <span slot="subtitle"
-              >by
-              <calcite-link href="">example_user</calcite-link>
-            </span>
-            <div>
-              Created: Apr 22, 2019
-              <br />
-              Updated: Dec 9, 2019
-              <br />
-              View Count: 0
-            </div>
-            <calcite-button
-              type="button"
-              slot="footer-leading"
-              color="neutral"
-              scale="s"
-              id="card-icon-test-1"
-              icon-start="check"
-            ></calcite-button>
-            <div slot="footer-trailing">
+
+        <calcite-card>
+          <img slot="thumbnail" alt="Sample image alt" src="https://placem.at/places?w=260&h=160&txt=0" />
+          <h3 slot="title">Portland Businesses</h3>
+          <span slot="subtitle"
+            >by
+            <calcite-link href="">example_user</calcite-link>
+          </span>
+          <div>
+            Created: Apr 22, 2019
+            <br />
+            Updated: Dec 9, 2019
+            <br />
+            View Count: 0
+          </div>
+          <calcite-button
+            type="button"
+            slot="footer-leading"
+            color="neutral"
+            scale="s"
+            id="card-icon-test-1"
+            icon-start="check"
+          ></calcite-button>
+          <div slot="footer-trailing">
+            <calcite-button type="button" scale="s" color="neutral" id="card-icon-test-2" icon-start="stairs">
+            </calcite-button>
+            <calcite-button type="button" scale="s" color="neutral" id="card-icon-test-3" icon-start="ellipsis">
+            </calcite-button>
+            <calcite-dropdown type="hover">
               <calcite-button
-                type="button"
+                id="card-icon-test-4"
+                slot="dropdown-trigger"
                 scale="s"
                 color="neutral"
-                id="card-icon-test-2"
-                icon-start="stairs"
-              ></calcite-button>
-              <calcite-button
-                type="button"
-                scale="s"
-                color="neutral"
-                id="card-icon-test-3"
-                icon-start="ellipsis"
-              ></calcite-button>
-              <calcite-dropdown type="hover">
-                <calcite-button
-                  id="card-icon-test-4"
-                  slot="dropdown-trigger"
-                  scale="s"
-                  color="neutral"
-                  icon-start="caret-down"
-                >
-                </calcite-button>
-                <calcite-dropdown-group selection-mode="none">
-                  <calcite-dropdown-item>View details</calcite-dropdown-item>
-                  <calcite-dropdown-item>Duplicate</calcite-dropdown-item>
-                  <calcite-dropdown-item>Delete</calcite-dropdown-item>
-                </calcite-dropdown-group>
-              </calcite-dropdown>
-            </div>
-          </calcite-card>
-        </calcite-tooltip-manager>
+                icon-start="caret-down"
+              >
+              </calcite-button>
+              <calcite-dropdown-group selection-mode="none">
+                <calcite-dropdown-item>View details</calcite-dropdown-item>
+                <calcite-dropdown-item>Duplicate</calcite-dropdown-item>
+                <calcite-dropdown-item>Delete</calcite-dropdown-item>
+              </calcite-dropdown-group>
+            </calcite-dropdown>
+          </div>
+        </calcite-card>
 
         <!-- button labels -->
         <calcite-tooltip label="bottom placed tooltip" placement="bottom-start" reference-element="card-icon-test-1"
           >My great tooltip example
         </calcite-tooltip>
-        <calcite-tooltip label="bottom placed tooltip" placement="bottom-start" reference-element="card-icon-test-2"
-          >Sharing level: 2
+        <calcite-tooltip label="bottom placed tooltip" placement="bottom-start" reference-element="card-icon-test-2">
+          Sharing level: 2
         </calcite-tooltip>
         <calcite-tooltip label="top end placed tooltip" placement="top-end" reference-element="card-icon-test-3"
           >More...
@@ -151,32 +140,30 @@
       <!-- selectable card -->
       <div class="child">
         <p>Card with two buttons</p>
-        <calcite-tooltip-manager>
-          <calcite-card selectable>
-            <img slot="thumbnail" alt="Sample image alt" src="https://placem.at/places?w=260&h=160&txt=0" />
-            <h3 slot="title">My great Workforce project that might wrap two lines</h3>
-            <span slot="subtitle">Johnathan Smith</span>
-            <span slot="footer-leading">Nov 25, 2018</span>
-            <div slot="footer-trailing">
-              <calcite-button
-                id="card-icon-test-6"
-                scale="s"
-                appearance="transparent"
-                color="neutral"
-                icon-start="check-circle-f"
-              >
-              </calcite-button>
-              <calcite-button
-                id="card-icon-test-7"
-                scale="s"
-                appearance="transparent"
-                color="neutral"
-                icon-start="x-circle-f"
-              >
-              </calcite-button>
-            </div>
-          </calcite-card>
-        </calcite-tooltip-manager>
+        <calcite-card selectable>
+          <img slot="thumbnail" alt="Sample image alt" src="https://placem.at/places?w=260&h=160&txt=0" />
+          <h3 slot="title">My great Workforce project that might wrap two lines</h3>
+          <span slot="subtitle">Johnathan Smith</span>
+          <span slot="footer-leading">Nov 25, 2018</span>
+          <div slot="footer-trailing">
+            <calcite-button
+              id="card-icon-test-6"
+              scale="s"
+              appearance="transparent"
+              color="neutral"
+              icon-start="check-circle-f"
+            >
+            </calcite-button>
+            <calcite-button
+              id="card-icon-test-7"
+              scale="s"
+              appearance="transparent"
+              color="neutral"
+              icon-start="x-circle-f"
+            >
+            </calcite-button>
+          </div>
+        </calcite-card>
       </div>
 
       <!-- selectable and already selected card -->
