+++ src/demos/link.html
@@ -166,11 +166,9 @@
       <div class="trigger-external-wrap-paragraph right-aligned-text">tooltip trigger</div>
 
       <div class="trigger-external-wrap-paragraph">
-        <calcite-tooltip-manager>
-          <calcite-link class="font-size-1" icon-start="information" id="tooltip-example-fourteen"
-            >tooltip trigger</calcite-link
-          >
-        </calcite-tooltip-manager>
+        <calcite-link class="font-size-1" icon-start="information" id="tooltip-example-fourteen"
+          >tooltip trigger
+        </calcite-link>
       </div>
     </div>
