+++ src/demos/notice.html
@@ -157,26 +157,13 @@
       <div class="child right-aligned-text">With actions and tooltip</div>
 
       <div class="child">
-        <calcite-tooltip-manager>
-          <calcite-notice icon="information" scale="s" width="full" active dismissible>
-            <div slot="title">Title lorem ispum</div>
-            <div slot="message">Body lorem ispum</div>
-            <calcite-action
-              id="retry-action-s"
-              scale="s"
-              slot="actions-end"
-              title="Retry"
-              icon="reset"
-            ></calcite-action>
-            <calcite-action
-              id="info-action-s"
-              scale="s"
-              slot="actions-end"
-              title="Get info"
-              icon="attachment"
-            ></calcite-action>
-          </calcite-notice>
-        </calcite-tooltip-manager>
+        <calcite-notice icon="information" scale="s" width="full" active dismissible>
+          <div slot="title">Title lorem ispum</div>
+          <div slot="message">Body lorem ispum</div>
+          <calcite-action id="retry-action-s" scale="s" slot="actions-end" title="Retry" icon="reset"></calcite-action>
+          <calcite-action id="info-action-s" scale="s" slot="actions-end" title="Get info" icon="attachment">
+          </calcite-action>
+        </calcite-notice>
 
         <calcite-tooltip reference-element="retry-action-s" placement="top">Retry that action</calcite-tooltip>
         <calcite-tooltip reference-element="info-action-s" placement="top">Attach a file</calcite-tooltip>
@@ -326,26 +313,13 @@
       <div class="child right-aligned-text">With actions and tooltip</div>
 
       <div class="child">
-        <calcite-tooltip-manager>
-          <calcite-notice icon="information" scale="m" width="full" active dismissible>
-            <div slot="title">Title lorem ispum</div>
-            <div slot="message">Body lorem ispum</div>
-            <calcite-action
-              id="retry-action-s"
-              scale="m"
-              slot="actions-end"
-              title="Retry"
-              icon="reset"
-            ></calcite-action>
-            <calcite-action
-              id="info-action-s"
-              scale="m"
-              slot="actions-end"
-              title="Get info"
-              icon="attachment"
-            ></calcite-action>
-          </calcite-notice>
-        </calcite-tooltip-manager>
+        <calcite-notice icon="information" scale="m" width="full" active dismissible>
+          <div slot="title">Title lorem ispum</div>
+          <div slot="message">Body lorem ispum</div>
+          <calcite-action id="retry-action-s" scale="m" slot="actions-end" title="Retry" icon="reset"></calcite-action>
+          <calcite-action id="info-action-s" scale="m" slot="actions-end" title="Get info" icon="attachment">
+          </calcite-action>
+        </calcite-notice>
 
         <calcite-tooltip reference-element="retry-action-s" placement="top">Retry that action</calcite-tooltip>
         <calcite-tooltip reference-element="info-action-s" placement="top">Attach a file</calcite-tooltip>
@@ -488,26 +462,13 @@
       <div class="child right-aligned-text">With actions and tooltip</div>
 
       <div class="child">
-        <calcite-tooltip-manager>
-          <calcite-notice icon="information" scale="l" width="full" active dismissible>
-            <div slot="title">Title lorem ispum</div>
-            <div slot="message">Body lorem ispum</div>
-            <calcite-action
-              id="retry-action-s"
-              scale="l"
-              slot="actions-end"
-              title="Retry"
-              icon="reset"
-            ></calcite-action>
-            <calcite-action
-              id="info-action-s"
-              scale="l"
-              slot="actions-end"
-              title="Get info"
-              icon="attachment"
-            ></calcite-action>
-          </calcite-notice>
-        </calcite-tooltip-manager>
+        <calcite-notice icon="information" scale="l" width="full" active dismissible>
+          <div slot="title">Title lorem ispum</div>
+          <div slot="message">Body lorem ispum</div>
+          <calcite-action id="retry-action-s" scale="l" slot="actions-end" title="Retry" icon="reset"></calcite-action>
+          <calcite-action id="info-action-s" scale="l" slot="actions-end" title="Get info" icon="attachment">
+          </calcite-action>
+        </calcite-notice>
 
         <calcite-tooltip reference-element="retry-action-s" placement="top">Retry that action</calcite-tooltip>
         <calcite-tooltip reference-element="info-action-s" placement="top">Attach a file</calcite-tooltip>
